# packages/gopeed-openapi/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Result_void_": {
				"description": "REST API result",
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "The response code, `0` means success, other values means error."
					},
					"msg": {
						"type": "string",
						"description": "The response message, if `code` != `0`, this field will contain error message."
					},
					"data": {
						"description": "The response data, if `code` == `0`, this field will contain response data."
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_string_": {
				"description": "REST API result",
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "The response code, `0` means success, other values means error."
					},
					"msg": {
						"type": "string",
						"description": "The response message, if `code` != `0`, this field will contain error message."
					},
					"data": {
						"type": "string",
						"description": "The response data, if `code` == `0`, this field will contain response data."
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InstallExtension": {
				"description": "Request body for installing an extension",
				"properties": {
					"url": {
						"type": "string",
						"description": "Git URL or local path (dev mode)"
					},
					"devMode": {
						"type": "boolean",
						"description": "Enable dev mode (load from local folder)"
					}
				},
				"required": [
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExtensionRepository": {
				"description": "Extension git repository info",
				"properties": {
					"url": {
						"type": "string",
						"description": "Git repository URL"
					},
					"directory": {
						"type": "string",
						"description": "Sub-directory within the repository (if extension is not at root)"
					}
				},
				"required": [
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExtensionMatch": {
				"description": "Extension script match rules",
				"properties": {
					"urls": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "URL match patterns (Chrome extension match pattern format)"
					},
					"labels": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Label keys to match on requests"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ExtensionScript": {
				"description": "Extension script definition",
				"properties": {
					"event": {
						"type": "string",
						"description": "Activation event name, e.g. \"onResolve\", \"onStart\", \"onError\", \"onDone\""
					},
					"match": {
						"$ref": "#/components/schemas/ExtensionMatch",
						"description": "Match rules"
					},
					"entry": {
						"type": "string",
						"description": "Entry JS file path (relative to extension directory)"
					}
				},
				"required": [
					"event",
					"entry"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SettingType": {
				"type": "string",
				"enum": [
					"string",
					"number",
					"boolean"
				],
				"description": "Extension setting type"
			},
			"ExtensionSettingOption": {
				"description": "Extension setting option (for select-type settings)",
				"properties": {
					"label": {
						"type": "string"
					},
					"value": {}
				},
				"required": [
					"label",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExtensionSetting": {
				"description": "Extension setting definition",
				"properties": {
					"name": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"required": {
						"type": "boolean"
					},
					"type": {
						"$ref": "#/components/schemas/SettingType"
					},
					"value": {},
					"options": {
						"items": {
							"$ref": "#/components/schemas/ExtensionSettingOption"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"title",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Extension": {
				"description": "Extension info",
				"properties": {
					"identity": {
						"type": "string",
						"description": "Global unique identity, format: \"author@name\""
					},
					"name": {
						"type": "string"
					},
					"author": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"version": {
						"type": "string",
						"description": "Semantic version, e.g. \"1.0.0\""
					},
					"homepage": {
						"type": "string"
					},
					"repository": {
						"$ref": "#/components/schemas/ExtensionRepository"
					},
					"scripts": {
						"items": {
							"$ref": "#/components/schemas/ExtensionScript"
						},
						"type": "array"
					},
					"settings": {
						"items": {
							"$ref": "#/components/schemas/ExtensionSetting"
						},
						"type": "array"
					},
					"disabled": {
						"type": "boolean",
						"description": "Whether the extension is disabled"
					},
					"devMode": {
						"type": "boolean",
						"description": "Whether the extension is running in dev mode"
					},
					"devPath": {
						"type": "string",
						"description": "Local path for dev mode extensions"
					},
					"createdAt": {
						"type": "string",
						"description": "Created time, ISO 8601 format"
					},
					"updatedAt": {
						"type": "string",
						"description": "Updated time, ISO 8601 format"
					}
				},
				"required": [
					"identity",
					"name",
					"title",
					"version"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_Extension-Array_": {
				"description": "REST API result",
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "The response code, `0` means success, other values means error."
					},
					"msg": {
						"type": "string",
						"description": "The response message, if `code` != `0`, this field will contain error message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Extension"
						},
						"type": "array",
						"description": "The response data, if `code` == `0`, this field will contain response data."
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_Extension_": {
				"description": "REST API result",
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "The response code, `0` means success, other values means error."
					},
					"msg": {
						"type": "string",
						"description": "The response message, if `code` != `0`, this field will contain error message."
					},
					"data": {
						"$ref": "#/components/schemas/Extension",
						"description": "The response data, if `code` == `0`, this field will contain response data."
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateExtensionSettings": {
				"description": "Request body for updating extension settings",
				"properties": {
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object",
						"description": "Settings key-value pairs"
					}
				},
				"required": [
					"settings"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SwitchExtension": {
				"description": "Request body for switching extension enabled/disabled",
				"properties": {
					"status": {
						"type": "boolean",
						"description": "true to enable, false to disable"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCheckExtensionResp": {
				"description": "Response body for checking extension update",
				"properties": {
					"newVersion": {
						"type": "string",
						"description": "New version string, empty if already up-to-date"
					}
				},
				"required": [
					"newVersion"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result_UpdateCheckExtensionResp_": {
				"description": "REST API result",
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "The response code, `0` means success, other values means error."
					},
					"msg": {
						"type": "string",
						"description": "The response message, if `code` != `0`, this field will contain error message."
					},
					"data": {
						"$ref": "#/components/schemas/UpdateCheckExtensionResp",
						"description": "The response data, if `code` == `0`, this field will contain response data."
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"X-Api-Token": {
				"type": "apiKey",
				"name": "X-Api-Token",
				"in": "header",
				"description": "Api Token for authentication"
			}
		}
	},
	"info": {
		"title": "Introduction",
		"version": "v1",
		"description": "This is gopeed RESTful API document. You can use this API to manage download tasks.\n\n# Response Format\n\nAll APIs return JSON format response. The response format is:\n\n```json\n{\n  \"code\": 0,\n  \"message\": \"\",\n  \"data\": null\n}\n```\n\n- `code`: The response code, `0` means success, other values means error.\n- `message`: The response message, if `code` != `0`, this field will contain error message.\n- `data`: The response data, if `code` == `0`, this field will contain response data.\n\n# Authentication\n\nWhen configed API token in Gopeed, should pass `X-Api-Token` in request header.\n\n# JavaScript Library\n\nWe provide a JavaScript library to help you use this API. You can install it by npm:\n\n```bash\nnpm install @gopeed/rest\n```\n\nThen you can use it in your code:\n\n```js\nimport Client from \"@gopeed/rest\";\n\n(async function(){\n  const client = new Client();\n  // Config host and API token\n  // const client = new Client({\"host:\":\"http://api.xxx.com\", \"token\": \"your api token\"});\n\n  // Create a http download task\n  await client.createTask({\n    \"req\":{\n      \"url\": \"https://gopeed.com\"\n    }\n  });\n\n  // Create a bt download task\n  await client.createTask({\n    \"req\":{\n      \"url\": \"magnet:?xt=urn:btih:xxx\"\n    }\n  });\n});\n````\n",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"url": "https://gopeed.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/extensions": {
			"post": {
				"operationId": "InstallExtension",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_string_"
								}
							}
						}
					}
				},
				"description": "Install an extension from a Git URL or a local folder (dev mode)",
				"summary": "Install extension",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InstallExtension"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetExtensions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_Extension-Array_"
								}
							}
						}
					}
				},
				"description": "Get all installed extensions",
				"summary": "Get extensions",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/extensions/{identity}": {
			"get": {
				"operationId": "GetExtension",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_Extension_"
								}
							}
						}
					}
				},
				"description": "Get a single extension by identity",
				"summary": "Get extension",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": [
					{
						"description": "- Extension identity (e.g. \"author@name\")",
						"in": "path",
						"name": "identity",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteExtension",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_void_"
								}
							}
						}
					}
				},
				"description": "Delete an extension",
				"summary": "Delete extension",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": [
					{
						"description": "- Extension identity",
						"in": "path",
						"name": "identity",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/extensions/{identity}/settings": {
			"put": {
				"operationId": "UpdateExtensionSettings",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_void_"
								}
							}
						}
					}
				},
				"description": "Update settings of an extension",
				"summary": "Update extension settings",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": [
					{
						"description": "- Extension identity",
						"in": "path",
						"name": "identity",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateExtensionSettings"
							}
						}
					}
				}
			}
		},
		"/api/v1/extensions/{identity}/switch": {
			"put": {
				"operationId": "SwitchExtension",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_void_"
								}
							}
						}
					}
				},
				"description": "Enable or disable an extension",
				"summary": "Switch extension",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": [
					{
						"description": "- Extension identity",
						"in": "path",
						"name": "identity",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SwitchExtension"
							}
						}
					}
				}
			}
		},
		"/api/v1/extensions/{identity}/update": {
			"get": {
				"operationId": "UpdateCheckExtension",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_UpdateCheckExtensionResp_"
								}
							}
						}
					}
				},
				"description": "Check if a new version is available for an extension",
				"summary": "Check extension update",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": [
					{
						"description": "- Extension identity",
						"in": "path",
						"name": "identity",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UpdateExtension",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result_void_"
								}
							}
						}
					}
				},
				"description": "Upgrade an extension to the latest version",
				"summary": "Update extension",
				"tags": [
					"Extensions"
				],
				"security": [
					{
						"X-Api-Token": []
					}
				],
				"parameters": [
					{
						"description": "- Extension identity",
						"in": "path",
						"name": "identity",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
...
}